requirements.txt
albumentations==1.3.0
gradio==3.39.0
matplotlib==3.7.2
mlflow==2.5.0
numpy==1.25.2
opencv_python_headless==4.8.0.74
pandas==1.4.4
Pillow==10.0.0
pydicom==2.1.2
pylibjpeg==1.2.0
scikit_image==0.18.2
torch==1.12.0
torchmetrics==0.11.4
torchvision==0.13.0
tqdm==4.46.0
